+++ rwlock6_t.c
@@ -42,11 +42,10 @@
  */
 
 #include "test.h"
-#include <sys/timeb.h>
 
 static pthread_rwlock_t rwlock1 = PTHREAD_RWLOCK_INITIALIZER;
 
-static int bankAccount = 0;
+static long bankAccount = 0;
 
 void * wrfunc(void * arg)
 {
@@ -60,7 +59,7 @@ void * wrfunc(void * arg)
 
 void * rdfunc(void * arg)
 {
-  int ba = -1;
+  long ba = -1;
   struct timespec abstime = { 0, 0 };
   struct _timeb currSysTime;
   const uint32_t NANOSEC_PER_MILLISEC = 1000000;
@@ -71,13 +70,13 @@ void * rdfunc(void * arg)
   abstime.tv_nsec = NANOSEC_PER_MILLISEC * currSysTime.millitm;
 
 
-  if ((int) arg == 1)
+  if ((long) arg == 1)
     {
       abstime.tv_sec += 1;
       assert(pthread_rwlock_timedrdlock(&rwlock1, &abstime) == ETIMEDOUT);
       ba = 0;
     }
-  else if ((int) arg == 2)
+  else if ((long) arg == 2)
     {
       abstime.tv_sec += 3;
       assert(pthread_rwlock_timedrdlock(&rwlock1, &abstime) == 0);
@@ -95,10 +94,10 @@ main()
   pthread_t wrt2;
   pthread_t rdt1;
   pthread_t rdt2;
-  int wr1Result = 0;
-  int wr2Result = 0;
-  int rd1Result = 0;
-  int rd2Result = 0;
+  long wr1Result = 0;
+  long wr2Result = 0;
+  long rd1Result = 0;
+  long rd2Result = 0;
 
   bankAccount = 0;
